+++ src/Exception/NumberFormatException.php
@@ -7,7 +7,7 @@ namespace Brick\Math\Exception;
 /**
  * Exception thrown when attempting to create a number from a string with an invalid format.
  */
-class NumberFormatException extends MathException
+final class NumberFormatException extends MathException
 {
     public static function invalidFormat(string $value) : self
     {