+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Minimum PHP version is now 8.2
 - `BigNumber` classes are now `readonly`
+- exception classes are now `final`
 
 ## [0.13.1](https://github.com/brick/math/releases/tag/0.13.1) - 2025-03-29